Rename current brief response manifests and questions as legacy
We rename `edit_brief_response` manifest to
`legacy_edit_brief_response` as we will shortly be deprecating this
flow. By renaming it to legacy we make room for a new manifest
to be created that will be called `edit_brief_response` as we don't
want to have to name it something awkward like `new_edit_brief_response`.
Similarly, we will need a legacy manifest for displaying a brief
response which can then be replaced with the new 'display_brief_response'
manifest. We won't be creating a legacy manifest for outputting a
brief response as although we need to maintain the legacy flow for it
for the foreseable future (as opposed to 2 weeks for the other manifests),
we will hack the CSV generation manually rather than driving legacy
content using a manifest.

Not only will the manifests be changing, but also the questions that
are asked related to a brief response so questions are also
prefixed with "legacy". To ensure the same data will be saved
to our API, we add "id"s to those without to ensure data isn't saved
under a different key (i.e. we don't want "dayRate" to now be saved
as "legacyDayRate").